pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.8",
    "groq>=0.16.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=3.4.1",
]
